Improve image rendering
